vide processing function

+++ preprocess/video_preprocess.py
@@ -89,7 +89,7 @@ def preprocess_raw_video(video_file_path, dim=72, plot=True, face_crop=True):
                 h = int(bounding_box.height * img.shape[0])
 
                 # cv2.rectangle(img, (x, int(y - 0.2 * h)), (x + w, y + h), (0, 255, 0), 2)
-                roi = img_as_float(img[int(y - 0.2 * h):y + 1.1 * h, x:x + w, :])
+                roi = img_as_float(img[int(y - 0.2 * h):int(y + 1.1 * h), x:x + w, :])
             vidLxL = cv2.resize(roi, (dim, dim), interpolation=cv2.INTER_AREA)
             prev_roi = roi
         else:
